pyproject.toml
[tool.poetry]
name = "mapumap"
version = "0.1.0"
description = ""
authors = ["Yongha Hwang <yongha.hwang@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
umap-learn = "^0.5.3"
pandas = "^2.0.2"
scikit-learn = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
